образец проекта полностью портирован, так же добавлен вспомогательный Point2f (был в javax.vecmath, не было у нас)

+++ src/test/java/com/secretsofgreenery/math/LinearAlgebraTest.java
@@ -1,6 +1,6 @@
 package com.secretsofgreenery.math;
 
-import main.java.com.example.math.LinearAlgebra;
+import com.secretsofgreenery.math.LinearAlgebra;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -10,13 +10,13 @@ class LinearAlgebraTest {
 
     @Test
     void testSolveSystem2x2() {
-        double[][] A = {
+        float[][] A = {
                 {2, 1},
                 {1, 3}
         };
-        double[] b = {5, 10};
+        float[] b = {5, 10};
 
-        double[] x = LinearAlgebra.solveSystem(A, b);
+        float[] x = LinearAlgebra.solveSystem(A, b);
 
         assertEquals(1, x[0], 1e-8);
         assertEquals(3, x[1], 1e-8);
@@ -27,16 +27,16 @@ class LinearAlgebraTest {
 
     @Test
     void testSolveSystem3x3() {
-        double[][] A = {
+        float[][] A = {
                 {1, 2, 3},
                 {4, 5, 6},
                 {7, 8, 10}
         };
-        double[] b = {6, 15, 25};
+        float[] b = {6, 15, 25};
 
-        double[] x = LinearAlgebra.solveSystem(A, b);
+        float[] x = LinearAlgebra.solveSystem(A, b);
 
-        double[] result = new double[3];
+        float[] result = new float[3];
         for (int i = 0; i < 3; i++) {
             for (int j = 0; j < 3; j++) {
                 result[i] += A[i][j] * x[j];
@@ -50,36 +50,36 @@ class LinearAlgebraTest {
 
     @Test
     void testSolveSystemInvalidMatrix() {
-        double[][] A = {
+        float[][] A = {
                 {1, 2},
                 {3, 4},
                 {5, 6}
         };
-        double[] b = {1, 2, 3};
+        float[] b = {1, 2, 3};
 
         assertThrows(IllegalArgumentException.class, () -> LinearAlgebra.solveSystem(A, b));
     }
 
     @Test
     void testSolveSystemSingular() {
-        double[][] A = {
+        float[][] A = {
                 {1, 2},
                 {2, 4}
         };
-        double[] b = {1, 2};
+        float[] b = {1, 2};
 
         assertThrows(IllegalArgumentException.class, () -> LinearAlgebra.solveSystem(A, b));
     }
 
     @Test
     void testSolveSystemWithZeroDiagonal() {
-        double[][] A = {
+        float[][] A = {
                 {0, 1},
                 {1, 0}
         };
-        double[] b = {2, 3};
+        float[] b = {2, 3};
 
-        double[] x = LinearAlgebra.solveSystem(A, b);
+        float[] x = LinearAlgebra.solveSystem(A, b);
 
         assertEquals(3, x[0], 1e-8);
         assertEquals(2, x[1], 1e-8);
@@ -87,15 +87,15 @@ class LinearAlgebraTest {
 
     @Test
     void testSolveSystem4x4() {
-        double[][] A = {
+        float[][] A = {
                 {2, 0, 0, 0},
                 {0, 3, 0, 0},
                 {0, 0, 4, 0},
                 {0, 0, 0, 5}
         };
-        double[] b = {6, 12, 20, 30};
+        float[] b = {6, 12, 20, 30};
 
-        double[] x = LinearAlgebra.solveSystem(A, b);
+        float[] x = LinearAlgebra.solveSystem(A, b);
 
         assertEquals(3, x[0], 1e-8);
         assertEquals(4, x[1], 1e-8);
